implemented the new reusable inputs and deleted old page-specific input jsx components
