+++ src/screens/App/screens/Seller/screens/Manage/screens/Product/components/utils/Pages/AddProduct.jsx
@@ -1,18 +1,13 @@
 import React, { useEffect } from "react";
 import { useForm } from "../../../../../../../../../../../hooks/useForm";
-import {
-  AreaInput,
-  DataListInput,
-  DefaultInput,
-  GalleryInput,
-  NameInput,
-  PriceInput,
-  SubmitCTA,
-} from "../ProductForm";
+
 import axios from "../../../../../../../../../../../shared/caller";
 import useAlert from "../../../../../../../../../../../hooks/useAlert";
 import { useHistory } from "react-router-dom";
 import { useManageProduct } from "../../../../../../../../../../../hooks/useManage";
+import { InputFirst } from "../../../../../../../../../../../shared/Components/input/InputBase";
+import { EmbossedInput } from "../../../../../../../../../../../shared/Components/input/Inputs";
+import Button from "../../../../../../../../../../../shared/Components/button/Button";
 
 function AddProduct() {
   const history = useHistory();
@@ -154,31 +149,45 @@ function AddProduct() {
   return (
     <div className="flex flex-col lg:flex-row gap-6 lg:gap-6 xl:gap-10 2xl:gap-20 justify-between">
       <div className=" w-full xl:w-3/5 lg:w-1/2 space-y-5 lg:space-y-7 xl:space-y-8">
-        <NameInput
+        <InputFirst
+          label="Product Name"
+          error={errors.item}
+          type="text"
           name="item"
           value={values.item}
-          error={errors.item}
           onChange={handleInput}
           placeholder="Name of your new product"
+          border="py-1.5 border-b-2 rounded-t shadow"
+          inputStyle="bg-transparent rounded-t-sm px-2 text-lg"
+          focusWithin="focus-within:border-my-accent"
         />
 
         <div
           className="flex w-full 
           flex-col xs:flex-row gap-4 xl:gap-6 "
         >
-          <PriceInput
-            label="Retail Price"
+          <EmbossedInput
+            type="number"
             name="retailPrice"
             value={values.retailPrice}
-            error={errors.retailPrice}
             onChange={handleInput}
             placeholder="Price in peso"
+            label="Retail Price"
+            error={errors.retailPrice}
             width="w-4/6 xs:w-1/2 sm:w-5/12 lg:w-3/5 xl:w-5/12"
+            icon={
+              <span
+                className="rounded bg-white px-1
+                text-my-accent font-semibold text-md "
+              >
+                PHP
+              </span>
+            }
           />
 
-          <DefaultInput
+          <EmbossedInput
             type="number"
-            label="Stock Onhand"
+            label="Onhand"
             name="inventory"
             value={values.inventory}
             error={errors.inventory}
@@ -192,7 +201,8 @@ function AddProduct() {
           className="flex w-full 
           flex-col xs:flex-row gap-4 xl:gap-6 "
         >
-          <PriceInput
+          <EmbossedInput
+            type="number"
             label="Wholesale Price"
             name="wholesalePrice"
             value={values.wholesalePrice}
@@ -200,9 +210,17 @@ function AddProduct() {
             onChange={handleInput}
             placeholder="Price in peso"
             width="w-4/6 xs:w-1/2 sm:w-5/12 lg:w-3/5 xl:w-5/12"
+            icon={
+              <span
+                className="rounded bg-white px-2 py-1.5
+                text-my-accent font-semibold text-md "
+              >
+                PHP
+              </span>
+            }
           />
 
-          <DefaultInput
+          <EmbossedInput
             type="number"
             label="Wholesale Cap"
             name="wholesaleCap"
@@ -214,16 +232,21 @@ function AddProduct() {
         </div>
 
         <div className="flex flex-col xs:flex-row md:flex-col xl:flex-row gap-4 xl:gap-6 w-full">
-          <DataListInput
+          <EmbossedInput
+            type="datalist"
             name="category"
+            label="Category"
             value={values.category}
             error={errors.category}
             onChange={handleInput}
             placeholder="Category of product"
             width="w-full xs:w-3/5 lg:w-full"
+            listId="category-list"
+            listData={[{ value: "Some Value", text: "Some value" }]}
           />
 
-          <DefaultInput
+          <EmbossedInput
+            type="text"
             label="Brand"
             name="brand"
             value={values.brand}
@@ -234,7 +257,8 @@ function AddProduct() {
           />
         </div>
 
-        <AreaInput
+        <EmbossedInput
+          type="textarea"
           label="Product Description"
           name="description"
           value={values.description}
@@ -243,7 +267,8 @@ function AddProduct() {
           placeholder="Description of the product"
         />
 
-        <DefaultInput
+        <EmbossedInput
+          type="text"
           label="Keywords"
           helper="seperate tags using a coma ',' "
           name="keywords"
@@ -264,9 +289,79 @@ function AddProduct() {
 
         <div className="border-b border-gray-300 rounded"></div>
 
-        <SubmitCTA isLoading={isLoading} onClick={handleFormSubmit} />
+        <Button
+          isLoading={isLoading}
+          onClick={handleFormSubmit}
+          buttonClass="bg-my-accent text-white px-8 sm:px-10 md:px-12 py-3 font-semibold text-sm uppercase 
+      transition duration-200 ease-linear rounded
+      hover:ring-2 hover:ring-my-accent hover:ring-opacity-70 hover:ring-offset-2
+      active:ring active:ring-my-accent active:ring-opacity-30 "
+          svgClass="text-white"
+        >
+          Upload Product
+        </Button>
       </div>
     </div>
   );
 }
 export default AddProduct;
+
+function GalleryInput({ name, value, error, onChange }) {
+  return (
+    <div className="space-y-2">
+      {/* preview of image */}
+      <div className="h-28 lg:h-36 xl:h-40 bg-my-white-tint bg-opacity-70 rounded flex items-center p-4">
+        {!value || value === "" ? (
+          <span className="font-medium text-base text-gray-500">
+            Preview of the image
+          </span>
+        ) : (
+          <img
+            alt="Image preview"
+            src={value}
+            className="object-contain h-full"
+          />
+        )}
+      </div>
+
+      <div className="flex flex-col xs:flex-row gap-4">
+        <label className="h-14 lg:h-16 w-14 lg:w-16 flex items-center justify-center cursor-pointer text-gray-800 bg-my-white-tint transition duration-200 ease-linear hover:bg-gray-200 hover:shadow active:text-my-accent">
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            className="h-7 lg:h-8 w-7 lg:w-8"
+            fill="none"
+            viewBox="0 0 24 24"
+            stroke="currentColor"
+          >
+            <path
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth={2}
+              d="M12 4v16m8-8H4"
+            />
+          </svg>
+
+          <input
+            type="file"
+            className="hidden"
+            accept="image"
+            // name={name}
+            // onChange={onChange}
+            // value={value}
+          />
+        </label>
+
+        <EmbossedInput
+          type="text"
+          label="Temporary"
+          name={name}
+          value={value}
+          error={error}
+          onChange={onChange}
+          width="w-full"
+          placeholder="Image address"
+        />
+      </div>
+    </div>
+  );
+}
